+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/cleaner/CleanerInitializer.java
@@ -12,6 +12,7 @@ import io.github.sakurawald.fuji.core.command.annotation.CommandRequirement;
 import io.github.sakurawald.fuji.core.config.handler.abst.BaseConfigurationHandler;
 import io.github.sakurawald.fuji.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.fuji.core.event.impl.ServerLifecycleEvents;
+import io.github.sakurawald.fuji.core.manager.Managers;
 import io.github.sakurawald.fuji.core.service.type_formatter.TypeFormatter;
 import io.github.sakurawald.fuji.core.document.annotation.ColorBox;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
@@ -148,7 +149,10 @@ public class CleanerInitializer extends ModuleInitializer {
 
     @Override
     protected void onInitialize() {
-        ServerLifecycleEvents.SERVER_STARTED.register(server -> new CleanerJob().schedule());
+        ServerLifecycleEvents.SERVER_STARTED.register(server -> {
+            CleanerJob cleanerJob = new CleanerJob();
+            Managers.getScheduleManager().scheduleJob(cleanerJob);
+        });
     }
 
 }