+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/fuji/FujiInitializer.java
@@ -8,7 +8,6 @@ import io.github.sakurawald.fuji.core.command.annotation.CommandRequirement;
 import io.github.sakurawald.fuji.core.command.annotation.CommandSource;
 import io.github.sakurawald.fuji.core.config.Configs;
 import io.github.sakurawald.fuji.core.document.inspection.CommandsInspectionGui;
-import io.github.sakurawald.fuji.core.job.abst.BaseJob;
 import io.github.sakurawald.fuji.core.manager.Managers;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.core.CoreInitializer;
@@ -49,7 +48,7 @@ public class FujiInitializer extends ModuleInitializer {
         Managers.getModuleManager().reloadModuleInitializers();
 
         // Reload jobs.
-        BaseJob.rescheduleAll();
+        Managers.getScheduleManager().rescheduleJobs();
 
         TextHelper.sendMessageByKey(source, "reload");
         return CommandHelper.Return.SUCCESS;