+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/tab/TabListInitializer.java
@@ -7,6 +7,7 @@ import io.github.sakurawald.fuji.core.auxiliary.minecraft.TextHelper;
 import io.github.sakurawald.fuji.core.config.handler.abst.BaseConfigurationHandler;
 import io.github.sakurawald.fuji.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.fuji.core.event.impl.ServerLifecycleEvents;
+import io.github.sakurawald.fuji.core.manager.Managers;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.tab.config.model.TabListConfigModel;
 import io.github.sakurawald.fuji.module.initializer.tab.job.RenderHeaderAndFooterJob;
@@ -35,7 +36,10 @@ public class TabListInitializer extends ModuleInitializer {
 
     @Override
     protected void onInitialize() {
-        ServerLifecycleEvents.SERVER_STARTED.register(server -> new RenderHeaderAndFooterJob().schedule());
+        ServerLifecycleEvents.SERVER_STARTED.register(server -> {
+            RenderHeaderAndFooterJob renderHeaderAndFooterJob = new RenderHeaderAndFooterJob();
+            Managers.getScheduleManager().scheduleJob(renderHeaderAndFooterJob);
+        });
     }
 
     @Override