+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [core] feature: attach detailed exception message for error log.
 - [fuji] feature: new `/fuji inspect jobs` command, to list all `registered jobs` and their `fire dates`.
 - [fuji] feature: attach `jobs` registered by this module in its module details GUI.
+- [fuji] feature: add document string for all jobs.